[FLINK-38458][table] Extract `SchemaConverter` and `ReferenceManager` to make them reusable
