+++ flink-table/flink-table-planner/src/main/java/org/apache/flink/table/planner/operations/converters/table/SqlAlterTableDropWatermarkConverter.java
@@ -1,0 +1,60 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.table.planner.operations.converters.table;
+
+import org.apache.flink.sql.parser.ddl.SqlAlterTableDropWatermark;
+import org.apache.flink.table.api.Schema;
+import org.apache.flink.table.api.ValidationException;
+import org.apache.flink.table.catalog.ResolvedCatalogTable;
+import org.apache.flink.table.catalog.TableChange;
+import org.apache.flink.table.operations.Operation;
+import org.apache.flink.table.planner.operations.converters.SchemaReferencesManager;
+
+import java.util.List;
+import java.util.function.Function;
+
+/** Convert ALTER TABLE DROP WATERMARK to generate an updated {@link Schema}. */
+public class SqlAlterTableDropWatermarkConverter
+        extends AbstractAlterTableConverter<SqlAlterTableDropWatermark> {
+    @Override
+    protected Operation convertToOperation(
+            SqlAlterTableDropWatermark dropWatermark,
+            ResolvedCatalogTable oldTable,
+            ConvertContext context) {
+        if (oldTable.getResolvedSchema().getWatermarkSpecs().isEmpty()) {
+            throw new ValidationException(
+                    String.format(
+                            "%sThe base table does not define any watermark strategy.",
+                            EX_MSG_PREFIX));
+        }
+
+        Schema.Builder schemaBuilder = Schema.newBuilder();
+        SchemaReferencesManager.buildUpdatedColumn(
+                schemaBuilder, oldTable, (builder, column) -> builder.fromColumns(List.of(column)));
+        SchemaReferencesManager.buildUpdatedPrimaryKey(
+                schemaBuilder, oldTable, Function.identity());
+
+        return buildAlterTableChangeOperation(
+                dropWatermark,
+                List.of(TableChange.dropWatermark()),
+                schemaBuilder.build(),
+                oldTable,
+                context.getCatalogManager());
+    }
+}
